requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Use pydantic-core from wheel - pin to a version with pre-built wheels
pydantic-core==2.14.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Handling
Pillow==11.2.1
imagehash==4.3.1
python-magic==0.4.27

# Email
resend==2.22.0
jinja2==3.1.6

# PDF Generation
weasyprint==60.1
reportlab==4.0.9

# QR Codes
qrcode[pil]==7.4.2

# HTTP Client
httpx==0.25.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Monitoring
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production Server
gunicorn==21.2.0

# Additional Dependencies
typing-extensions==4.8.0
email-validator==2.1.0
bcrypt==4.0.1
cffi==1.16.0
pycparser==2.21
ecdsa==0.18.0
pyasn1==0.5.0
rsa==4.9
six==1.16.0